Backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib
bcrypt
paho-mqtt
geopy
python-dotenv
pydantic-settings
python-multipart
sqlmodel